🦄fixes🪲fix bridge page
