+++ views/bridge/PolygonBridge.jsx
@@ -256,7 +256,7 @@ const PolygonBridge = () => {
 
   return (
     <>
-      <Stack gap={3} className={`bg-light-subtle p-3 rounded-4 ${""}`}>
+      <Stack gap={3} className={`bg-secondary p-3 rounded-4 ${""}`}>
         <Stack
           gap={2}
           direction="horizontal"
@@ -299,7 +299,7 @@ const PolygonBridge = () => {
             type="number"
             placeholder="Enter Amount"
             name="ethAmount"
-            className="caption2 flex-grow-1 bg-transparent"
+            className="caption2 flex-grow-1 bg-transparent border border-0"
             style={{ outline: "none" }}
             value={displayInputAmountValue}
             onChange={handleAmountOnChange}
